# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "金丝雀蜜罐 CanaryGuard AntiFraud",
    "description": "蜜罐反诈双形态主控：REST API 同一内核，WebUI/MCP 均为客户端（D1）。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/accounts/{url_name}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Account",
        "description": "查询已建档账号画像。",
        "operationId": "get_account_api_v1_accounts__url_name__get",
        "parameters": [
          {
            "name": "url_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Url Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Account Api V1 Accounts  Url Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/account-bridges": {
      "get": {
        "tags": [
          "account-bridges"
        ],
        "summary": "List Matrix",
        "description": "平台矩阵：7 平台调研结论（名称/AI 可接入/接入方式/开源项目/状态），不含任何明文凭据。",
        "operationId": "list_matrix_api_v1_account_bridges_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Matrix Api V1 Account Bridges Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account-bridges/{platform}": {
      "get": {
        "tags": [
          "account-bridges"
        ],
        "summary": "Get Detail",
        "description": "单平台详情：矩阵字段 + config_fields（动态表单） + has_config（不含明文）。",
        "operationId": "get_detail_api_v1_account_bridges__platform__get",
        "parameters": [
          {
            "name": "platform",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Platform"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Detail Api V1 Account Bridges  Platform  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "account-bridges"
        ],
        "summary": "Delete Config",
        "description": "删除配置：清空 config_enc、status=unconfigured、写事件 + gate_logs 审计（凭据不可恢复）。",
        "operationId": "delete_config_api_v1_account_bridges__platform__delete",
        "parameters": [
          {
            "name": "platform",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Platform"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DeleteConfigRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Config Api V1 Account Bridges  Platform  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account-bridges/{platform}/config": {
      "put": {
        "tags": [
          "account-bridges"
        ],
        "summary": "Save Config",
        "description": "保存配置：敏感字段（cookie/token/password/secret）Fernet 双层加密落库，status=configured，\n写 events(account_bridge_configured) + gate_logs 审计。",
        "operationId": "save_config_api_v1_account_bridges__platform__config_put",
        "parameters": [
          {
            "name": "platform",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Platform"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Save Config Api V1 Account Bridges  Platform  Config Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account-bridges/{platform}/test": {
      "post": {
        "tags": [
          "account-bridges"
        ],
        "summary": "Test",
        "description": "测试连接：结构校验 + URL/token 的 http GET 连通（timeout 5s）；cookie 仅解密+解析。\n返回 {ok, detail, latency_ms}；失败 status='error'。",
        "operationId": "test_api_v1_account_bridges__platform__test_post",
        "parameters": [
          {
            "name": "platform",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Platform"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TestRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Test Api V1 Account Bridges  Platform  Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account-bridges/{platform}/ingest": {
      "post": {
        "tags": [
          "account-bridges"
        ],
        "summary": "Ingest Dm",
        "description": "私信接入核心 webhook：桥接器（WeChatFerry/OneBot/长轮询等）推送私信 →\ndetections(pending) + account_intel upsert + events(dm_received) → 反向侦察流水线。",
        "operationId": "ingest_dm_api_v1_account_bridges__platform__ingest_post",
        "parameters": [
          {
            "name": "platform",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Platform"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestDMRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Ingest Dm Api V1 Account Bridges  Platform  Ingest Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account-bridges/{platform}/agent-import": {
      "get": {
        "tags": [
          "account-bridges"
        ],
        "summary": "Agent Import",
        "description": "本地 agent（DSH/Claude Code/Codex）一键导入配置：MCP streamable-http + 环境变量 + 桥脚本提示。",
        "operationId": "agent_import_api_v1_account_bridges__platform__agent_import_get",
        "parameters": [
          {
            "name": "platform",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Platform"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Agent Import Api V1 Account Bridges  Platform  Agent Import Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteConfigRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 200,
            "title": "Reason",
            "description": "审计理由（可选）",
            "default": "删除账号桥接配置"
          }
        },
        "type": "object",
        "title": "DeleteConfigRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestDMRequest": {
        "properties": {
          "from": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "From",
            "description": "对方账号标识（知乎 url_token / 微信昵称 / QQ 号 / TG 用户名等）"
          },
          "text": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Text",
            "description": "私信原文"
          },
          "dm_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Dm Id",
            "description": "桥接器侧消息 id（去重用，可选）"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload",
            "description": "桥接器附加载荷（原样存入事件，可选）"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "IngestDMRequest",
        "description": "私信接入 webhook body（桥接器调用）。字段名 from 为 Python 关键字，用别名。"
      },
...
      "SaveConfigRequest": {
        "properties": {
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Fields",
            "description": "平台配置字段（config_fields 白名单内）"
          },
          "reason": {
            "type": "string",
            "maxLength": 200,
            "title": "Reason",
            "description": "审计理由（gate_logs）",
            "default": "配置账号桥接"
          }
        },
        "type": "object",
        "required": [
          "fields"
        ],
        "title": "SaveConfigRequest"
      },
...
      "TestRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 200,
            "title": "Reason",
            "description": "审计理由（可选）",
            "default": "测试账号桥接"
          }
        },
        "type": "object",
        "title": "TestRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}